X::name(): new virtual method
The API change is compatible, the app source compiles without any
changes. However the ABI has change is not exactly compatible.
Fortunately the ABI change is not that bad, and the binary linked
with the previous version of the library runs with the new one
(despite the dynamic linker warning).